+++ changelog.md
@@ -5,3 +5,9 @@ All notable changes to `macmotp/locale` will be documented in this file
 ## v0.1.0 - 2024-09-20
 
 - initial release
+
+## v0.1.1 - 2024-09-20
+
+- added currencies and languages as objects
+- added flag property
+- added translations of capitals